feat: add GitHub Release creation script and related documentation
This commit introduces a new feature allowing GitHub releases to be automatically created. A bash script (entrypoint.sh) has been added which captures input parameters and prepares `gh release create` command to generate a new release on GitHub.

Additionally, comprehensive documentation on how to use this feature has been added to the README.md file. The action.yml file describes the inputs details for the GitHub Action to create a new release. This feature will be invaluable for automation and streamlining of software releases on GitHub.