refactor: update `entrypoint.sh` for readability and modularity
Changes are made in the entrypoint.sh script to make it more readable and modular. A new function 'add_argument' has been introduced to modularize the appending of arguments to the 'arguments' array, reducing code repetition. The checks for command-line inputs have been reorganized to make the code easier to understand. The process of adding release details (title, notes, etc.) to the arguments array has also been updated to use the new 'add_argument' function. This refactor aims to make the code easier to manage and understand.