ci: update actions/checkout version in GitHub workflow
The version of actions/checkout used in the release workflow has been updated from v2 to v3.5.3. This was done to ensure we are using the latest stable release of the action which includes important bug fixes and enhancements. The environment variable for GITHUB_TOKEN has also been removed as it is no longer needed.