ci: refactor GitHub Actions workflows
The "checks.yml" workflow was deleted and its ShellCheck job was integrated into the "prepare.yml" workflow. This was done to streamline the workflows and ensure shell scripts are checked for syntax errors before drafting a release. Also, the README was updated to reflect the changes in the workflow structure.