refactor: update handling of GitHub token in entrypoint.sh
Modified how the GITHUB_TOKEN is set and used in the entrypoint.sh script. This was done to improve the security and maintainability of the code. Instead of unsetting and later authenticating with gh auth, the GITHUB_TOKEN is now directly exported as an environment variable. This approach reduces unnecessary lines of code and potential points of failure.