ci: update GitHub release action reference in workflow
The GitHub action used to create releases in our CI/CD workflow has been updated. We've moved from using version 1.0.8 of the 'nekofar/create-github-release' action to using the 'develop' branch of the same action. This change intends to take advantage of the latest improvements and updates made in this action, enhancing our release processes.