fix: remove export from `GITHUB_TOKEN` setting
The export keyword was removed while setting the `GITHUB_TOKEN` environment variable in `entrypoint.sh`. This is because the export was unnecessary - variables declared in the script will be automatically available to any subprocesses. Removing this keyword simplifies the code and avoids potential confusion or misuse of the `export` command.