refactor: trim input token and authenticate GitHub CLI
Changes were made in the entrypoint.sh to improve the token handling. Instead of setting the GitHub token as an environment variable, the token is now trimmed for any white spaces for accuracy, and then is used to authenticate with the GitHub CLI. This provides an improved level of security and accuracy of token usage.