test: add tests for input validation in entrypoint.bats
Added two test cases in a newly created file, entrypoint.bats. The first test ensures that the 'tag' input is present, the second test verifies the INPUT_TOKEN to be a valid GitHub token. These tests increase the reliability of our code by ensuring that the correct input is being used in the script.